chore: cleanup events
