+++ packages/ripple/src/compiler/phases/2-analyze/index.js
@@ -4,7 +4,6 @@ import { create_scopes, ScopeRoot } from '../../scope.js';
 import {
   get_delegated_event,
   is_element_dom_element,
-  is_event_attribute,
   is_inside_component,
   is_ripple_import,
   is_void_element,
@@ -13,6 +12,7 @@ import { extract_paths } from '../../../utils/ast.js';
 import is_reference from 'is-reference';
 import { prune_css } from './prune.js';
 import { error } from '../../errors.js';
+import { is_event_attribute } from '../../../utils/events';
 
 function mark_control_flow_has_template(path) {
   for (let i = path.length - 1; i >= 0; i -= 1) {
