+++ packages/ripple/src/runtime/internal/client/render.js
@@ -1,12 +1,8 @@
 import { destroy_block, ref } from './blocks';
 import { REF_PROP } from './constants';
-import {
-  get_descriptors,
-  get_own_property_symbols,
-  get_prototype_of,
-  is_event_handler,
-  get_event_name,
-} from './utils';
+import { get_descriptors, get_own_property_symbols, get_prototype_of } from './utils';
+import { event } from './events';
+import { get_attribute_event_name, is_delegated, is_event_attribute } from '../../../utils/events';
 
 export function set_text(text, value) {
   // For objects, we apply string coercion (which might make things like $state array references in the template reactive) before diffing
@@ -75,11 +71,17 @@ export function set_attributes(element, attributes) {
 
     if (key === 'class') {
       set_class(element, value);
-    } else if (is_event_handler(key, value)) {
+    } else if (is_event_attribute(key)) {
       // Handle event handlers in spread props
-      const event_name = get_event_name(key);
-      // Set up event delegation by storing the handler on the element
-      element['__' + event_name] = value;
+      const event_name = get_attribute_event_name(key);
+
+      if (is_delegated(event_name)) {
+        // Use delegation for delegated events
+        element['__' + event_name] = value;
+      } else {
+        // Use addEventListener for non-delegated events
+        event(event_name, element, value);
+      }
     } else {
       set_attribute(element, key, value);
     }
